Move WASAPI endpoint to endpoint module
